Moved various mimir plates into yggdrasil
